Rename variable and paths
`OUTPUT_DATA_ROOT_DIR` → `OUTPUT_DATA_DIR`

We also change the path to include a hyphen between the name of the
tokenizer and the "tok" suffix (`my-tiny-c4-gpt2tok` →
`my-tiny-c4-gpt2-tok`).

The "root" naming is not appropriate because the data needs to be merged
before it can be used anyway.