Fix patching NeMo dataset for correct merging
Previously, the document indices would not be updated. It's a bit sad
that we have to create an entire semi-duplicate of the method for this
but there's no way around it unless we re-load the index file (this
could be achieved efficiently by adding an LRU cache, but that seems
like more complexity than checking the implementation from time to
time).