Remove undesirable setting
It was not about sharding as the documentation implied, but about a
buffer size after which a flush would occur. Having this buffer size
unlimited means that we would just run out of memory at some point. The
default value of 1 << 26 = 67108864 is fine even with 128 processe per
node (requires approx. 8.6 GB RAM).